chrome/content/mochitest/tests/test_matcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Filter matcher tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">

  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    prepareFilterComponents();

    function compareShortcuts(text, expected)
    {
      expected.push(null);

      for each (let filter in [Filter.fromText(text), Filter.fromText("@@" + text)])
      {
        let result = [];
        for each (let shortcut in expected)
        {
          result.push(matcher.findShortcut(filter.text));
          if (shortcut)
            matcher.add(Filter.fromText(shortcut));
        }
  
        is(result.join("\n"), expected.join("\n"), "Shortcut candidates for " + filter.text);
        matcher.clear();
      }
    }
    function checkMatch(filters, location, contentType, docDomain, thirdParty, expected)
    {
      for each (let filter in filters)
        matcher.add(Filter.fromText(filter));

      let result = matcher.matchesAny(location, contentType, docDomain, thirdParty);
      if (result)
        result = result.text;

      is(result, expected, "match(" + location + ", " + contentType + ", " + docDomain + ", " + (thirdParty ? "third-party" : "first-party") + ") with:\n" + filters.join("\n"));
      matcher.clear();
    }
    function cacheCheck(location, contentType, docDomain, thirdParty, expected)
    {
      let result = matcher.matchesAny(location, contentType, docDomain, thirdParty);
      if (result)
        result = result.text;

      is(result, expected, "match(" + location + ", " + contentType + ", " + docDomain + ", " + (thirdParty ? "third-party" : "first-party") + ") with static filters");
    }

    is(typeof Matcher, "function", "typeof Matcher");
    is(typeof blacklistMatcher, "object", "typeof blacklistMatcher");
    is(typeof whitelistMatcher, "object", "typeof whitelistMatcher");
    ok(blacklistMatcher instanceof Matcher, "blacklistMatcher is a Matcher instance");
    ok(whitelistMatcher instanceof Matcher, "whitelistMatcher is a Matcher instance");

    let matcher = new Matcher();

    compareShortcuts("*", []);
    compareShortcuts("asdf", []);
    compareShortcuts("asdf1234", ["asdf1234"]);
    compareShortcuts("asdf12345", ["asdf1234", "sdf12345"]);
    compareShortcuts("asdf123456", ["sdf12345", "df123456", "asdf1234"]);
    compareShortcuts("asdf123456a", ["sdf12345", "df123456", "asdf1234", "f123456a"]);
    compareShortcuts("asdf123456as", ["df123456", "f123456a", "sdf12345", "123456as", "asdf1234"]);
    compareShortcuts("asdf*1234", []);
    compareShortcuts("asdf1234*5678", ["asdf1234"]);
    compareShortcuts("asdf12345*678", ["sdf12345", "asdf1234"]);
    compareShortcuts("asdf1234*d*qwer5678", ["asdf1234", "qwer5678"]);
    compareShortcuts("asdf^1234", []);
    compareShortcuts("asdf1234^5678", ["asdf1234"]);
    compareShortcuts("asdf12345^678", ["sdf12345", "asdf1234"]);
    compareShortcuts("asdf1234^d^qwer5678", ["asdf1234", "qwer5678"]);
    compareShortcuts("|asdf1234", ["asdf1234"]);
    compareShortcuts("asdf1234|", ["asdf1234"]);
    compareShortcuts("|asdf12345|", ["asdf1234", "sdf12345"]);
    compareShortcuts("asdf|1234", ["asdf|123", "sdf|1234"]);
    compareShortcuts("||asdf1234", ["asdf1234"]);
    compareShortcuts("||asdf12345|", ["asdf1234", "sdf12345"]);
    compareShortcuts("/foo/", []);
    compareShortcuts("/foobarbar/", []);
    compareShortcuts("*/foo/", []);
    compareShortcuts("*/foobarbar/", ["foobarba", "oobarbar", "/foobarb", "obarbar/"]);
    compareShortcuts("^/foobarbar/", ["foobarba", "oobarbar", "/foobarb", "obarbar/"]);

    checkMatch([], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["abc"], "http://abc/def", "IMAGE", null, false, "abc");
    checkMatch(["abc", "ddd"], "http://abc/def", "IMAGE", null, false, "abc");
    checkMatch(["ddd", "abc"], "http://abc/def", "IMAGE", null, false, "abc");
    checkMatch(["ddd", "abd"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["abc", "://abc/d"], "http://abc/def", "IMAGE", null, false, "://abc/d");
    checkMatch(["://abc/d", "abc"], "http://abc/def", "IMAGE", null, false, "://abc/d");
    checkMatch(["|http://"], "http://abc/def", "IMAGE", null, false, "|http://");
    checkMatch(["|http://abc"], "http://abc/def", "IMAGE", null, false, "|http://abc");
    checkMatch(["|abc"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["|/abc/def"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["/def|"], "http://abc/def", "IMAGE", null, false, "/def|");
    checkMatch(["/abc/def|"], "http://abc/def", "IMAGE", null, false, "/abc/def|");
    checkMatch(["/abc/|"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["http://abc/|"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["|http://abc/def|"], "http://abc/def", "IMAGE", null, false, "|http://abc/def|");
    checkMatch(["|/abc/def|"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["|http://abc/|"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["|/abc/|"], "http://abc/def", "IMAGE", null, false, null);
    checkMatch(["||example.com/abc"], "http://example.com/abc/def", "IMAGE", null, false, "||example.com/abc");
    checkMatch(["||com/abc/def"], "http://example.com/abc/def", "IMAGE", null, false, "||com/abc/def");
    checkMatch(["||com/abc"], "http://example.com/abc/def", "IMAGE", null, false, "||com/abc");
    checkMatch(["||mple.com/abc"], "http://example.com/abc/def", "IMAGE", null, false, null);
    checkMatch(["||.com/abc/def"], "http://example.com/abc/def", "IMAGE", null, false, null);
    checkMatch(["||http://example.com/"], "http://example.com/abc/def", "IMAGE", null, false, null);
    checkMatch(["||example.com/abc/def|"], "http://example.com/abc/def", "IMAGE", null, false, "||example.com/abc/def|");
    checkMatch(["||com/abc/def|"], "http://example.com/abc/def", "IMAGE", null, false, "||com/abc/def|");
    checkMatch(["||example.com/abc|"], "http://example.com/abc/def", "IMAGE", null, false, null);
    checkMatch(["abc", "://abc/d", "asdf1234"], "http://abc/def", "IMAGE", null, false, "://abc/d");
    checkMatch(["foo*://abc/d", "foo*//abc/de", "://abc/de", "asdf1234"], "http://abc/def", "IMAGE", null, false, "://abc/de");
    checkMatch(["abc$third-party", "abc$~third-party", "ddd"], "http://abc/def", "IMAGE", null, false, "abc$~third-party");
    checkMatch(["abc$third-party", "abc$~third-party", "ddd"], "http://abc/def", "IMAGE", null, true, "abc$third-party");
    checkMatch(["//abc/def$third-party", "//abc/def$~third-party", "//abc_def"], "http://abc/def", "IMAGE", null, false, "//abc/def$~third-party");
    checkMatch(["//abc/def$third-party", "//abc/def$~third-party", "//abc_def"], "http://abc/def", "IMAGE", null, true, "//abc/def$third-party");
    checkMatch(["abc$third-party", "abc$~third-party", "//abc/def"], "http://abc/def", "IMAGE", null, true, "//abc/def");
    checkMatch(["//abc/def", "abc$third-party", "abc$~third-party"], "http://abc/def", "IMAGE", null, true, "//abc/def");
    checkMatch(["abc$third-party", "abc$~third-party", "//abc/def$third-party"], "http://abc/def", "IMAGE", null, true, "//abc/def$third-party");
    checkMatch(["abc$third-party", "abc$~third-party", "//abc/def$third-party"], "http://abc/def", "IMAGE", null, false, "abc$~third-party");
    checkMatch(["abc$third-party", "abc$~third-party", "//abc/def$~third-party"], "http://abc/def", "IMAGE", null, true, "abc$third-party");
    checkMatch(["abc$image", "abc$script", "abc$~image"], "http://abc/def", "IMAGE", null, false, "abc$image");
    checkMatch(["abc$image", "abc$script", "abc$~script"], "http://abc/def", "SCRIPT", null, false, "abc$script");
    checkMatch(["abc$image", "abc$script", "abc$~image"], "http://abc/def", "OTHER", null, false, "abc$~image");
    checkMatch(["//abc/def$image", "//abc/def$script", "//abc/def$~image"], "http://abc/def", "IMAGE", null, false, "//abc/def$image");
    checkMatch(["//abc/def$image", "//abc/def$script", "//abc/def$~script"], "http://abc/def", "SCRIPT", null, false, "//abc/def$script");
    checkMatch(["//abc/def$image", "//abc/def$script", "//abc/def$~image"], "http://abc/def", "OTHER", null, false, "//abc/def$~image");
    checkMatch(["abc$image", "abc$~image", "//abc/def"], "http://abc/def", "IMAGE", null, false, "//abc/def");
    checkMatch(["//abc/def", "abc$image", "abc$~image"], "http://abc/def", "IMAGE", null, false, "//abc/def");
    checkMatch(["abc$image", "abc$~image", "//abc/def$image"], "http://abc/def", "IMAGE", null, false, "//abc/def$image");
    checkMatch(["abc$image", "abc$~image", "//abc/def$script"], "http://abc/def", "IMAGE", null, false, "abc$image");
    checkMatch(["abc$domain=foo.com", "abc$domain=bar.com", "abc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "foo.com", false, "abc$domain=foo.com");
    checkMatch(["abc$domain=foo.com", "abc$domain=bar.com", "abc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "bar.com", false, "abc$domain=bar.com");
    checkMatch(["abc$domain=foo.com", "abc$domain=bar.com", "abc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "baz.com", false, "abc$domain=~foo.com|~bar.com");
    checkMatch(["abc$domain=foo.com", "cba$domain=bar.com", "ccc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "foo.com", false, "abc$domain=foo.com");
    checkMatch(["abc$domain=foo.com", "cba$domain=bar.com", "ccc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "bar.com", false, null);
    checkMatch(["abc$domain=foo.com", "cba$domain=bar.com", "ccc$domain=~foo.com|~bar.com"], "http://abc/def", "IMAGE", "baz.com", false, null);
    checkMatch(["abc$domain=foo.com", "cba$domain=bar.com", "ccc$domain=~foo.com|~bar.com"], "http://ccc/def", "IMAGE", "baz.com", false, "ccc$domain=~foo.com|~bar.com");

    // Testing whether result cache messes up results
    matcher.add(Filter.fromText("abc$image"));
    matcher.add(Filter.fromText("abc$script"));
    matcher.add(Filter.fromText("abc$~image,~script,~document"));
    matcher.add(Filter.fromText("cba$third-party"));
    matcher.add(Filter.fromText("cba$~third-party,~script"));
    matcher.add(Filter.fromText("http://def$image"));
    matcher.add(Filter.fromText("http://def$script"));
    matcher.add(Filter.fromText("http://def$~image,~script,~document"));
    matcher.add(Filter.fromText("http://fed$third-party"));
    matcher.add(Filter.fromText("http://fed$~third-party,~script"));
    cacheCheck("http://abc", "IMAGE", null, false, "abc$image");
    cacheCheck("http://abc", "SCRIPT", null, false, "abc$script");
    cacheCheck("http://abc", "OTHER", null, false, "abc$~image,~script,~document");
    cacheCheck("http://cba", "IMAGE", null, false, "cba$~third-party,~script");
    cacheCheck("http://cba", "IMAGE", null, true, "cba$third-party");
    cacheCheck("http://def", "IMAGE", null, false, "http://def$image");
    cacheCheck("http://def", "SCRIPT", null, false, "http://def$script");
    cacheCheck("http://def", "OTHER", null, false, "http://def$~image,~script,~document");
    cacheCheck("http://fed", "IMAGE", null, false, "http://fed$~third-party,~script");
    cacheCheck("http://fed", "IMAGE", null, true, "http://fed$third-party");
    cacheCheck("http://abc_cba", "DOCUMENT", null, false, "cba$~third-party,~script");
    cacheCheck("http://abc_cba", "DOCUMENT", null, true, "cba$third-party");
    cacheCheck("http://abc_cba", "SCRIPT", null, false, "abc$script");
    cacheCheck("http://def?http://fed", "DOCUMENT", null, false, "http://fed$~third-party,~script");
    cacheCheck("http://def?http://fed", "DOCUMENT", null, true, "http://fed$third-party");
    cacheCheck("http://def?http://fed", "SCRIPT", null, false, "http://def$script");
  </script>
  </pre>
</body>
</html>
